+++ src/generic-table/interfaces/gt-row-meta.ts
@@ -1,0 +1,4 @@
+export interface GtRowMeta {
+  isOpen?:boolean;
+  isSelected?:boolean;
+}