+++ src/generic-table/interfaces/gt-row.ts
@@ -1,3 +1,4 @@
 export interface GtRow {
   isOpen?: boolean;
+  $$gtRowId?:any;
 }